(PUP-4201) Add support for structured logging
This commit removes the responsibility for formatting the log messages
from the IssueReporter and the ExternalFilerError classes. Instead,
the formatting is deferred and performed by the Log.to_s method (if
such formatting is at all needed).

Some instance variables were added, both to the ExternalFileError class
and to the Log class to accomodate detailed information about
issue_code, environment, node, and position.

A new Log destination 'json_file' was added. It will not use the
formatted message but instead output all details in a JSON style
Map. Exception backtrace is output as a JSON Array.