(PUP-4231) Log file content hash when using ensure=>present
When `ensure => present` and file content is provided via `content` or
`source`, the message logged was that the file was created. This is in
contrast to when `ensure => file` with content provided via `content` or
`source`, in which scenario the message logged contains the file's hash.
This makes both values log the file's hash.