(PUP-4230) Log content hash when removing a file
When removing a file with `ensure => absent`, the message logged is that
the file was removed. By contrast, when files are created or modified,
the message logged contains a hash of current and, if applicable,
previous contents. This changes file removal log entries to also contain
the hash of the removed file.