(PUP-4194) Move `logdir` default to /var/log/puppetlabs/puppet
Previously, puppet would set its `logdir` permissions to `root:root:750`
if run on a host without the `puppet` service user and group. If the
`puppetserver` package was installed later, it would create the `puppet`
user and group accounts, try to start, but fail to write to its log
directory `/var/log/puppetlabs/puppetserver`, because of the parent
directory permissions.

This commit moves puppet's logdir to `/var/log/puppetlabs/puppet`,
isolated from both `puppetserver` and `puppetmaster` packages, the latter
being the passenger-based compatibility packages.

This commit affects both the `puppetdlog` setting, which is pretty much
dead, and the puppet master webrick `masterhttplog` setting, which is
soon to be deprecated.