(PUP-4090) Zypper provider improvements
Prior to this commit, our zypper provider didn't work. When installing a
package with zypper, we pass in the `--name` flag, which is not
available in earlier versions of zypper that we account for. It also
breaks when you use `install_options`. It previously dropped in the
extra install options between `--name` and the package name to be
installed. This cuased it to fail.

This pull request reorders the options that we pass to the zypper
command. If the user has passed in the global option `--no-gpg-check`,
we pull that flag out, and pass it in where zypper expects it (after
`zypper` and before `install`).