(PUP-1072) fix streaming method
When streaming HTTP content by passing a block to the request method,
it's necessary to call an alternate method such as `request_get`
instead of `get` (in Net::HTTP). When following redirects, the code
needs to make sure that only the final response yields to the block.
To this end, always use HEAD requests, until one returns a non-redirect.
If a block was given, it is passed to the `request_xxx` method that
corresponds to the requested HTTP method. Otherwise, the method
is invoked directly and the result returned.