Parse /etc/filesystems with code instead of regex
The regex worked, but any comments or blank lines in the middle of a
stanza would end the stanza parsing resulting in a malformatted file.
Plus it was not easy to read or debug.

This updates the "parser" to use simple state when parsing the file and
does not barf on comments or blank lines in the middle of stanzas
(though they will be moved to the end of the stanza to avoid edge
cases on file regeneration).