(PUP-1279) ADSI Explicitly set user/group deleted
Previously, flush would only check against whether the
user/group variable was set before deciding to commit. This is
erroneous as the variable could be set (and sometimes is) by other
means. Explicitly set a flag when a user/group is deleted if the
user/group variable is set.

Without this change there are cases when the user/group variable will
be set and flush will fail, throwing an error even though the deletion
is successful.