(PUP-4340) Fix intermittent failure in test case
The source_attribute test was modified to wait for time to pass
sufficient that touching source files on the master would cause puppet
agent runs to resync the files. However, the test only checked
mtime/ctime on the master - which is unrelated to ctime/mtime of the
target files on the agents - and the agent target files are what will be
compared to the ctime/mtime of the source file to see if they need to be
resynced.

We could fix this by iterating over the agents - or assuming they'll
always be done in order and only querying the last agent - to check the
ctime/mtime stats of the target files using `puppet resource file
<filename>`, but that's more complicated and unlikely to be
significantly less than the 1 second sleep we used to have.

Revert to the previous behavior of waiting 1 second to touch the files
and try resyncing.