(PUP-4356) Remove puppetversion from PMT metadata
Before this commit, there was an undocumented field, puppetversion,
which was tracked by the Puppet::Module class.  An error would be thrown
if the metadata value was present and was not an exact string match to
Puppet.version.  This commit removes the recognition and checking of
that metadata key:value pair.

Note: Accessors and validation for puppetversion still exist in order to
not cause a breaking change, but they return nil.