(PUP-4356) Remove puppetversion from PMT metadata
Before this commit, there was an undocumented key, puppetversion, which
was tracked by the Puppet::Module class. This did nothing but throw an
error if the metadata value was an exact match to Puppet.version. This
commit removes the recognition and checking of that metadata key:value
pair.

Note: Accessors and validation for puppetversion are now separate and
ineffective, though don't cause breaking behavior change.