(PUP-4373) Windows User groups_insync? Compare SID
Use SIDS when comparing user groups. Implement similar behavior as
Group members when comparing whether the groups a user is in are in
sync with current. Without this change casing and names must match
exactly and can produce erroneous results.

With this change, the comparison will be by SIDs so it will allow for
the following when a user is a member of Administrators and Users
groups:

```
user { 'tim':
  ensure => present,
  groups => ['BUILTIN\Administrators','users'],
}
```