(PUP-4379) Fix interpolation of shortform var with _ continued with [
Before this, it was not possible to use the short form interpolation
of a variable starting with _ if continued with '[' or '.'. This is
caused by the fact that names cannot start with '_' and the
transformation of interpolated expressions did instead receive the
bare word '_x' as a LiteralString (which cannot be transformed to
a variable).

The fix modifies the lexer so that the first token is always
transformed if it is a NAME or a WORD, and if it is followed by
'[' or '.'.(Logic already in place handles the case where there is
only one expression being interpolated.