(PUP-4378) Improve output of label for collector objects
Under certain circumstances, a Collector object may be assigned to
a variable or placed in an Array or Hash. When transormed to a string
for human output, the output showed the full classname and id.

This changes the output to be a label "Catalog-Collector[type]", or
"Exported-Collector[type]", where type is the type being collected
e.g. "CatalogColloector[Notify]".