(PUP-4410) decrease build targets
With the switch to puppet 4, we are no longer building and shipping
individual package. Instead, we're shipping puppet as a part of the
puppet-agent package. This means we can reduce the number of cow and
mock build targets to only LTS releases (for community members who are
using the automation in the packaging repo to build puppet). We can also
stop building the dmg and the msi, as puppet will be delivered to those
platforms also with puppet-agent.