(PUP-2455) Fix logic in ticket_2455 acceptance test
Prior to this commit, the ticket_2455 acceptance test
fails if there are no Solaris agents.

With no valid agents, the interaction of 'skip_test'
after the setup of 'with_puppet_running_on' causes
the whole test to fail.

This commit updates the acceptance test to skip the setup
if there are no Solaris agents.