(PUP-2080) Make puppet function loading work
Puppet function loading required work after the rebase.

A fix was made (earlier) that makes module loaders reference the lib dir
instead of the parent of lib dir. This was done earlier because puppet
itself may not be "mounted" under a 'lib' directory. That fix made it
difficult to treat all relative paths the same way.

To overcome this (and related problems), this commit:

* Makes it possible for a loader to return which kinds of things it
  is set up to load (e.g. puppet system does not contain any .pp logic)
* The PuppetSmartPath understands that a loader's path is to its
  (possibly non existing) 'lib' directory and adjusts the "root"
  accordingly.
* A subclass of Closure called NamedClosure adds the ability to
  a closure to know its name - this is used for errors only. This
  avoids errors refering to "lambda", when the closure is a named
  function.

In order to make it simpler to create a function that represents a
puppet function the Closure class was modified to also be directly
invoked the same way as a Dispatch. This removes the need to have
an additional Dispatch just to translate `invoke` to `call`.

These changes made it possible to simplyfy logic.
Tests are added, and the loaders_spec is slightly reorganized to make
it easier to navigate (fixes poor example titles).