(maint) Use Puppet's chmod to set inital file mode
Previously, the test failed on Windows, because the test used ruby's File.chmod
to set the initial file mode. However, this is largely meaningless on Windows,
as it only sets or clears the read-only attribute for a file.

When puppet retrieves the current file mode, it is reported as 200644, since
there are permissions in the Windows access control list that can't be
represented in a POSIX mode.

This commit modifies the test to use puppet's chmod so that the initial state
is correct.