(PUP-4404) Use current user for security specs
Previously the security spec tests expected an Administrator user to exist
that would be used for testing generic owner and group manipulation. The
Administrator user is not guaranteed to exist, so use the current user
instead.

The spec tests set owner to :guest, then expect changing back to
:current_user to work. If #set_owner is broken then ownership will never
change away from :current_user and the test will pass. Verify that
set_owner to :guest works to avoid this case.