(PUP-4193) Make lambda a case option that is called as a predicate
Before this it was possible to do deep matching, but not combining
several values from the input in a predicate match.

This commit adds the ability to use a lambda as a case option. When
a match is performed, the test-expression is splatted as arguments
in a call to the lambda, and the result is used to determine if
a match was made or not.

notice case [2,1] {
 |$x, $y| { $x  == $y } : { 'equal' }
 |$x, $y| { $x  <  $y } : { 'ascending' }
 |$x, $y| { $x  >  $y } : { 'descending' }
}

Will notice 'descending'