(PUP-2894) Add multi assignment of variables in array
Before this it was not possible to directly assign multiple variables
from values in an array.:

This commit adds this ability so that:
```
[$a, $b, $c] = [1,2,3]
```
assigns the values 1,2,3 to $a, $b, $c respectively.

Nested arrays are supported:
```
[$a, [$b, $c]] = [1,[2,3]]
```

And if RHS is not an array, it is treated as an array of 1 value.

The numbers of variables and values must be the same.