(PUP-4394) Improve erroring for WEBrick server when Puppet user doesn't exist
The puppet-agent package does not create the puppet user or group
by default, meaning that attempting to run the webrick server without
manually creating the user results in Puppet failing hard.

This commit updates Puppet's webrick code to first check it the
puppet user exists, and only attempt to switch to it if it does.
Otherwise, we emit an error and exit. Previously, Puppet would
actually try to switch to the non-existent user and throw several
errors. We now only throw one.