(PUP-4404) Fix SID lookup when running as SYSTEM
When running as the SYSTEM account, doing lookup of the current user via
the WIN32OLE interface fails, because SYSTEM isn't recognized as a user.

We already have a test that uses the SYSTEM SID. That should be
sufficient, so remove the troublesome test.

Resolved by still removing the test even though it had changed between
the two branches.