(PUP-3988) Remove unused hostname checking feature
Before this, the validator had the ability to pass a single-feature
name when checking hostnames. This was used to check the parent of a
node. This is no longer needed since having a parent is now illegal.

This removes this feature. Tests that checked validation of parent
host name matches were removed a long time ago.