(PUP-3814) Only log compilation errors once
Before this all compilation errors were logged both by the compiler, and
then again by the apply application.

This splits up the error handling so that those errors that are logged
and processed by the compiler are caught and then causing an immediate
exit. Other errors (such as errors after the compiler has finished
compiling, or before being in a state where errors are rescued and
logged) are caught by the general logic at the end of the compile/apply
block.

This fix is not a guarantee there will never be duplicate output since
different compilation logic can be used (and it may not behave the same
way).