(PUP-4443) Add splat assign of multiple variables from a hash
Before this it was not possible to assign variables in a multi
assignment from a hash (only from an array). This feature is wanted
because it is of value to be able to compose the data to assign and this
is much easier to do when data is in hash form with keys reflecting the
variable name.

This commit makes it possible to assign variables in a multi assignment
from a hash such that:

[$a, $b] = { a => 1, b => 2}

assigns $a the value 1, and $b the value 2.