(PUP-4044) Curl repo config directly to SUT
Previously, we copied the puppetlabs-release package from
{yum,apt}.puppetlabs.com, the per-commit repo_config from
builds.puppetlabs.lan or nightlies.puppetlabs.com, and the actual repos
from either PC1, main or devel to the coordinator node, which then
scp'ed the files to each SUT.

This commit modifies the `install_repos_on` method to curl the
repo_config for the specified SHA to the SUT's system location e.g.
/etc/yum.repos.d. Once installed the repo_config specifies the URL from
which to install the package for the specified SHA, e.g

    baseurl=http://builds.puppetlabs.lan/puppet-agent/<SHA>/repos/el/7/PC1/x86_64/

This commit also skips installing the puppetlabs-release package.
In puppet 3.x, the package is required, because it provides dependencies
for puppet, e.g. facter, hiera. In the AIO world, this is no longer
necessary.