(PUP-4447) Give better error message for assignment to name and title
Before this, an attempt to assign to $name o $title, when not in a
lambda where $name or $title is a parameter would result in an error
saying "Cannot reassign variable ...".

This was bad because users does not understand that $name and $title are
always assigned in the body of a class or a define, and that top-scope
is really in the body of Class['main'].

This commit improves the error in this situation to read:
"Cannot reassign built in (or already assigned) variable '$...'"

The vagueness "built in (or already assigned)" is used since when the
error occurs it is difficult to compute if the evaluated logic is in the
body of a class or define or not. Instead, it gives this alternative
output when the name of the variable is "name" or "title". The only
downside is the potential confusion if:

* the parameter name $name or $title is used in a lambda AND
* an attempt is made to reassign those variables

This should be an aceptable trade off vs. the performance penalty it
would mean to pass along a more comprehensive context for the sake of
only differentiating from where the assignment was made.