(PUP-4233) Override is_puppetserver in passenger node definitions
Previously, the aio options.Rb file assumed we were running
puppetserver, which won't work for passenger.

This commit continues to use puppetserver by default for AIO
acceptance. However, it allows the passenger host definitions to
override this. The logic for switching between puppetserver and
passenger (where apache is either httpd or apache2) is then moved into
a single place. This allows us to retain the same `rake ci:test:aio`
task, and have the host definition completely specify which server to
use.

Note due to BKR-219, `is_puppetserver: false` must be specified in the
global CONFIG section, not the master specific settings.