(PUP-4234) Install passenger packages
Previously, we always installed `puppetserver` packages on the master.

This commit modifies the presuite step to install either the
`puppetserver` or `puppet-master-passenger` package on the master
depending on whether `is_puppetserver` is true or not. By default, we
will install 5.0.6.0 of the passenger package, and the version (tag or
sha) can be overridden `PASSENGER_VERSION` environment variable. Eventually,
that should be changed to 'nightly'.