(PUP-4374) Make *=> work in resource override expressions
Before this it was not possible to use a *=> to set resource override
attributes from a hash. This failed because the *=> evaluation results
in a nested array that the evaluation of the resource override
expression was not prepared to handle. This led to a somewhat mysterious
error.

The fix, now correctly finds the parent resource expression (this stops
the strange error from being reported). A subsequent problem was then
fixed where *=> results in an array of parameters inside an array. The
is simply to flatten the result (this is not based on input from the
user, so it is completely safe).