(PUP-4234) Configure passenger post-install
Prior to 4.0, the debian-specific passenger package would create SSL
keys and certs, update the apache configuration, and enable the
puppet-passenger site. These steps are no longer being done in 4.0
passenger packages, and we didn't provide a redhat passenger package until
now.

This commit configures passenger for debian and redat post-install. On both
systems, we need to rename the SSL keys and cert file names in the apache
configuration to match where puppet will later create them. Note we don't
actually create SSL keys and certs until the 040_ValidateSignCert step that
comes later.

On debian, we also need to enable specific modules and the puppet-passenger
site.

We also work around a packaging issue that installs config.ru in the
wrong place.