(PUP-3814) Add test that assert that errors are only printed once
It was difficult to assert that something was not printed twice at the
same time as it is asserted that it is printed once. Instead two
separate checks were made.

The problem is due to the difficulty of writing negative matches.It is
relatively easy to use negative look ahead with ?! in a regep, but then
combining it with matching the beginning of the output where there may
be coloration and other leading output was to hard to figure out.