(PUP-4463) Make Regexp Type without pattern use '' when used as pattern
Before this, if a Regexp without a pattern was used as a regular
expression, it would present a nil pattern - which for the split
function means use a default split on whitespace (other methods on
string may give it different meaning).

This makes a Regexp without a pattern behave as if the pattern was
// when asking for the regexp's compiled regular expression.

This also updates the functions split, and regsubst to either use the
type's compiled regexp, or the same rule (pattern || '').

As a type, the Regexp still represents all possible patterns.