(maint) Remove or-equals
Previously, the or-equals operator was applied to PASSENGER_VERSION and
SERVER_VERSION, which had the side effect of setting the environment
variable if not set.

This commit changes them to just use the `or` operator, as it's less
magical. The environment variables are not used anywhere else in
acceptance, so it should be safe to remove the side-effect.