(PUP-2630) Warn when node parameter is not updated
Prior to this commit, if merge attempted to set a node parameter
that already existed, it would quietly fail to update the value.
This meant that if a client fact value was being used for one of
the server side facts, it would not be overwritten with the correct
value and the user would never know.

Since we cannot change the behavior without breaking API, simply
add a warning so that the user will be aware if merge failed to
change the value of a node parameter.