(PUP-4499) Retain size_type for common type of two String types
The type calculator incorrectly assumes that the size_type of two PStringType
instances should be dropped when computing a common type. This commit changes
this so that the size_type will be the common type of the size_types of the
given string types provided that both have a size_type defined.