(PUP-1931) Improvements when mount options not specified
Currently, when a user specified a resource to be mounted, puppet allows
the user to not pass in any information for either fstype or options.
This was fine if there was an existing entry for that resource defined
in fstab, but if this was for a new entry then it would create an
invalid entry. Then when mount was called against the fstab file, mount
would barf everywhere.

This commit updates how we deal with updating the fstab file:

* Options omitted
  - Creating a new entry:
    * linux; `:options => 'defaults'`
    * non-linux; raise Puppet::Error
  - Updating an existing entry:
    * Existing entry options set to 'something,valid'; `:options => 'something,valid'`
    * Existing entry options not set
      - linux; `:options => 'defaults'`
      - non-linux; raise Puppet::Error
* fstype omitted
  - Creating a new entry; raise Puppet::Error
  - Updating an existing entry:
    * Existing entry fstype set to 'ext3'; `fstype => 'ext3'`
    * Existing entry fstype not set; raise Puppet::Error