(maint) Do not allow options or fstype to be an empty string
Currently, we allow the user to pass in an empty string for options like
fstype and options. If the user does this, Puppet will create an invalid
entry in the fstab file. We don't want this to happen, since mount will
barf if we allow the invalid entry to be written