(PUP-4420) Ensure pluginfacts don't ignore source permissions
Prior to this commit, the :source_permissions attribute was
not being properly propagated into the Posix MetaStat class,
and so source permissions were ignored for all resources.

External facts are required to use source permissions so that
the facts are executable when they are downloaded onto
the agent. This means that due to this regression, external
facts were being pluginsync'd to the client without execution
permissions.

This commit re-enables using source permissions for fact
pluginsync, while continuing to ignore them otherwise.

See commit 21edcd6 for the first fix for of a similar, earlier
regression.