(maint) Fix/update AppVeyor settings
The appveyor script gives the appearance of setting to Ruby 2.1.5,
but that particular folder doesn't actually exist. When the build runs
it should share the environment variables, the Gemfile.lock file and
the Ruby version it is set on. Additionally we should test against
the x64 architecture, which should catch a reasonable amount of
failures related to Windows. Adding a second architecture slows
the build down quite a bit (to more than double the amount of time),
so we won't go the route of adding a matrix for now.