(PUP-4520) Fix regression with case default option not being done last
Before this, and after 3.7.5, a case option being a literal default
was selected when encountered instead of being selected last if no other
option was selected.

This was introduced when adding either support for deep merge, or for
migration checking. There was no test finding this regression.
(A similar problem was found earlier with selector).

The fix was to remove the call to is_match? for the default option and
instead return false (to continue with other case option values in the
same option.

As a consequence, one migration test needed to be changed as it reported
a "type diff" for default vs. Resource type (which was actually wrong,
since a comparisson against default is not considered a type match diff)