(PUP-735) Ensure reports are finalized when catalog.apply fails
After the initial work on PUP-735, we were left with a situation where
certain catalog apply errors (including resource cycles) could casue a
report to not be finalized. An easy symptom of this to see is that the
metrics section of the report is empty.

This fixes the configurer to always finalize the report, and adds a
test.