(PUP-4527) Add MigrationChecker method to provide available migrations
Befoe this, reflection had to be used to see what a version of a
MigrationChecker implements.

This is now changed so all versions of puppet after 3.8.0 has a method
that returns the migration checks that are implemented. This makes it
possible to have newer versions of the migration checker support older
versions of puppet (>= 3.8.0).