(PUP-4527) Remove migration strategies and make it an error to call them
Before this, if missing to remove calls to no longer supported migration
checking methods it would silently just waste cycles. This can happen
when changes on an older branch is merged and calls to the
migration_checker for unsupported migration strategies are brought
forward (they should be removed).

After this, such calls will raise errors in the null/base implementation
of the migration checker.