(PUP-4436) Ensure ancestor directories are created
Prior to 4.0, puppet's `confdir` and `vardir` were `/etc/puppet` and
`/var/lib/puppet`, respectively, the parent directories always existed,
and all other file/directory settings were contained within. Since the
parent directories always existed, you could install puppet as a gem,
and run it as root, and puppet could create its `confdir`, etc.

This was also true when running puppet as non-root as `confdir` was
`~/.puppet` and `vardir` was `$confdir/var`.

In Puppet 4, `confdir` and `vardir` were changed to
`/etc/puppetlabs/puppet` and `/opt/puppetlabs/puppet/cache`, and their
ancestor directories are not guaranteed to exist.

Commit fccd76a6c3 tried to fix the issue by ensuring ancestor
directories were created in the per-user run mode. However, it didn't
cover the case when puppet is installed as a gem and run as root.

Also it created the ancestor directories as a side-effect of
resolving them. And it turns out that puppet always resolves the
default locations for both root and user directories, and later
overrides the default locations with values specified on the command
line, and resolves whether to use system or user run mode.

This commit moves the logic for creating ancestor directories to later
in the process after we know our run_mode, but before we try to apply
any settings catalogs, as that is when puppet will try to create its
`confdir`, etc.

An alternative approach would be to create puppet settings for the
parent directories for the purposes of having puppet create them when
applying the settings catalog. But it seems less desirable to create
essentially bogus settings that users shouldn't set, and the property
values would be different for root vs user.